redid ui

+++ app/src/main/java/com/ncnf/event/PublicEvent.java
@@ -19,10 +19,10 @@ public class PublicEvent extends Event {
 
     private List<Tag> tags;
 
-    private int price;
+    private double price;
     private int minAge;
 
-    public PublicEvent(String ownerId, String name, Date date, GeoPoint location, String address, String description, Type type, int minAge, int price) {
+    public PublicEvent(String ownerId, String name, Date date, GeoPoint location, String address, String description, Type type, int minAge, double price) {
         super(ownerId, name, date, location, address, type, Event.Visibility.PUBLIC, description);
 
         checkConstraints(minAge, price);
@@ -32,7 +32,7 @@ public class PublicEvent extends Event {
         this.price = price;
     }
 
-    public PublicEvent(String ownerId, UUID uuid, String name, Date date, GeoPoint location, String address, String description, Type type, List<String> attendees, int minAge, int price, List<Tag> tags) {
+    public PublicEvent(String ownerId, UUID uuid, String name, Date date, GeoPoint location, String address, String description, Type type, List<String> attendees, int minAge, double price, List<Tag> tags) {
         super(ownerId, uuid, name, date, location, address, type, Visibility.PUBLIC, attendees, description);
 
         checkConstraints(minAge, price);
@@ -42,14 +42,14 @@ public class PublicEvent extends Event {
         this.price = price;
     }
 
-    private void checkConstraints(int minAge, int price){
+    private void checkConstraints(int minAge, double price){
         if(!(minAge >= MIN_AGE && minAge <= MAX_AGE) || price < 0) {
             throw new IllegalArgumentException();
         }
     }
 
     public int getMinAge() { return minAge; }
-    public int getPrice() { return price; }
+    public double getPrice() { return price; }
     public List<Tag> getTags() { return new ArrayList<Tag>(tags); }
 
     public void setMinAge(int minAge) {
@@ -57,7 +57,7 @@ public class PublicEvent extends Event {
         this.minAge = minAge;
     }
 
-    public void setPrice(int price) { this.price = price; }
+    public void setPrice(double price) { this.price = price; }
 
     public void setTags(List<Tag> tags) {
         this.tags = new ArrayList<Tag>(tags);
